Removed all remaining uses of 64-bit types
Also removed include statements of stdint.h, as by #66.

Closes #66.

+++ walls-of-doom/clock.h
@@ -1,9 +1,7 @@
 #ifndef CLOCK_H
 #define CLOCK_H
 
-#include <stdint.h>
-
-typedef uint32_t Milliseconds;
+typedef long Milliseconds;
 
 /**
  * Returns a number of milliseconds.
